[feature/handle-keys-have-changed-86epu3nfx] Move logic to WrongKeyPopup component (from Main)
